+++ src/components/datepicker/DatePickerInput.js
@@ -78,9 +78,9 @@ class DatePickerInput extends PureComponent {
 
   render() {
     const {
-      bordered,
       className,
       dayPickerProps,
+      footer,
       inverse,
       inputProps,
       locale,
@@ -110,7 +110,7 @@ class DatePickerInput extends PureComponent {
           active={isPopoverActive}
           anchorEl={popoverAnchorEl}
           backdrop="transparent"
-          fullWidth
+          maxWidth="min-content"
           onEscKeyDown={this.handlePopoverClose}
           onOverlayClick={this.handlePopoverClose}
           position="end"
@@ -120,7 +120,6 @@ class DatePickerInput extends PureComponent {
         >
           <Box overflowY="auto">
             <DatePicker
-              bordered={bordered}
               className={datePickerClassNames}
               locale={locale}
               localeUtils={LocaleUtils}
@@ -131,6 +130,21 @@ class DatePickerInput extends PureComponent {
               {...dayPickerProps}
             />
           </Box>
+          {footer && (
+            <Box
+              backgroundColor="neutral"
+              backgroundTint="light"
+              borderColor="neutral"
+              borderTint="normal"
+              borderTopWidth={1}
+              flex="1 0 auto"
+              overflowX="hidden"
+              paddingHorizontal={3}
+              paddingVertical={3}
+            >
+              {footer}
+            </Box>
+          )}
         </Popover>
       </Box>
     );
@@ -138,12 +152,12 @@ class DatePickerInput extends PureComponent {
 }
 
 DatePickerInput.propTypes = {
-  /** If true we give a border to our wrapper. */
-  bordered: PropTypes.bool,
   /** A class name for the wrapper to give custom styles. */
   className: PropTypes.string,
   /** Object with props for the DatePicker component. */
   dayPickerProps: PropTypes.object,
+  /** A footer component, rendered at the bottom of the date picker */
+  footer: PropTypes.any,
   /** A custom function to format a date. */
   formatDate: PropTypes.func,
   /** Object with props for the Input component. */
@@ -167,7 +181,6 @@ DatePickerInput.propTypes = {
 };
 
 DatePickerInput.defaultProps = {
-  bordered: false,
   dayPickerProps: {},
   inputProps: {},
   inverse: false,